確実な文字列を挿入する部分をファイル分割した。
