+++ Solver/ConfirmSolver/InsertConfirmParts.h
@@ -1,0 +1,9 @@
+#ifndef INSERT_COMFIRM_PARTS
+#define INSERT_COMFIRM_PARTS
+
+#include "../../DataStructure/InputStructure/input_v1.h"
+#include "../../DataStructure/MiddleStructure/middle_predict_v1.h"
+
+void InsertConfirmParts(input_structure*, middle_predict_structure*);   // 確実に1つに定まる部分文字列を当てはめる。
+
+#endif
