+++ Solver/ConfirmSolver/Makefile
@@ -1,2 +1,2 @@
  test : test.c
-	gcc test.c SolveCertainString.c sort.c InitializeMiddleStructure.c ../../Generator/ReadDataFile.c ../../Settings/confirm_solver_setting.c
+	gcc test.c SolveCertainString.c sort.c InitializeMiddleStructure.c InsertConfirmParts.c ../../Generator/ReadDataFile.c ../../Settings/confirm_solver_setting.c
